refactor: 에로 로직을 JS에서 처리하도록 수정
